auth

+++ apps/api/user/user.service.go
@@ -8,7 +8,7 @@ import (
 )
 
 type UserService interface {
-	GetExistingUser(id int, ctx context.Context) (User, error)
+	GetExistingUser(id int64, ctx context.Context) (User, error)
 	CreateUser(user User, ctx context.Context) (sql.Result, error)
 }
 
@@ -28,6 +28,6 @@ func (s *service) CreateUser(user User, ctx context.Context) (sql.Result, error)
 	return s.repository.CreateUser(user, ctx)
 }
 
-func (s *service) GetExistingUser(id int, ctx context.Context) (User, error) {
+func (s *service) GetExistingUser(id int64, ctx context.Context) (User, error) {
 	return s.repository.GetExistingUser(id, ctx)
 }
